Rename feature to `discussion-filters`
